Move `AJAX` class to WooCommerce integration
